Pinpoints where the bug (id: 7)  was occurring
Ave Nero! This bug got me tilted so bad, I really had no idea if it was corrupting my data or not. Pretty hard to pinpoint, if I didn't had my subs and functions well defined it would had took forever. It motivated to create a documentation for the Spreadsheet changes.