Implements solution to bug (id: 7)
I found out the bug was happening because when it sorts the table it messes up with the cells original references and sh*t was stacking up, so it had a bunch of stacked up cells references somewhere in the Excel code within the spreadsheet file. So I fixed it by adding an statement to clean up the references every time it sorts it self. Next heavy code change I will remove the comments describing and tracking the problem and create a documentation with the changes.